+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Update module ([https://github.com/GotCms/GotCms/issues/113](https://github.com/GotCms/GotCms/issues/113))
 - Add module.config.php for custom modules ([https://github.com/GotCms/GotCms/issues/105](https://github.com/GotCms/GotCms/issues/105))
 - Bug ActivityLog Module ([https://github.com/GotCms/GotCms/issues/116](https://github.com/GotCms/GotCms/issues/116))
+- Check if data have changed during form edition ([https://github.com/GotCms/GotCms/issues/119](https://github.com/GotCms/GotCms/issues/119))
 
 ### Potential Breakage
 All acl have been changed, and if you added new roles, you must redefined them after the update.